pythoncode/requirements.txt
pyserial
pyaudio
numpy
matplotlib